+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [0.0.3-alpha] - 2018-10-14
+### Added
+- Use sprites/icons for types and pokemon in details
+- Added converters to customize details
+
+### Changed
+- Image of main window in [README.md](README.md)
+
 ## [0.0.2-alpha] - 2018-10-14
 ### Added
 - Datagrid events (context menu, filtering)
@@ -18,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Refactoring
 - [README.md](README.md)
 
-## [0.0.1-alpha] - 2018-10-14
+## [0.0.1-alpha] - 2018-10-13
 ### Added
 
 - The readme file, which uses [this](https://gist.github.com/PurpleBooth/109311bb0361f32d87a2) template
@@ -27,4 +35,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 [0.0.1-alpha]: https://github.com/nlabiris/pokemonstats/compare/v0.0.1...HEAD
-[0.0.2-alpha]: https://github.com/nlabiris/pokemonstats/compare/v0.0.1...HEAD
+[0.0.2-alpha]: https://github.com/nlabiris/pokemonstats/compare/v0.0.2...HEAD
+[0.0.3-alpha]: https://github.com/nlabiris/pokemonstats/compare/v0.0.3...HEAD